+++ djangocms_link/fields_select2.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 from django import forms
 
 from cms.models import Page
@@ -6,7 +5,7 @@ from cms.models import Page
 from django_select2.forms import ModelSelect2Widget
 
 
-class Select2PageSearchFieldMixin(object):
+class Select2PageSearchFieldMixin:
     search_fields = [
         'title_set__title__icontains',
         'title_set__menu_title__icontains',
@@ -17,10 +16,23 @@ class Select2PageSearchFieldMixin(object):
 class Select2PageSelectWidget(Select2PageSearchFieldMixin, ModelSelect2Widget):
     site = None
 
+    # show entries when clicking on it
+    def build_attrs(self, base_attrs, extra_attrs=None):
+        default_attrs = {"data-minimum-input-length": 0}
+        default_attrs.update(base_attrs)
+        attrs = super().build_attrs(default_attrs, extra_attrs=extra_attrs)
+        return attrs
+
     def get_queryset(self):
         if self.site:
             return Page.objects.drafts().on_site(self.site)
         return Page.objects.drafts()
+
+    # we need to implement jQuery ourselves, see #180
+    class Media:
+        js = (
+            "https://code.jquery.com/jquery-3.5.1.slim.min.js",
+        )
 
 
 class Select2PageSearchField(forms.ModelChoiceField):
@@ -28,4 +40,4 @@ class Select2PageSearchField(forms.ModelChoiceField):
 
     def __init__(self, *args, **kwargs):
         kwargs['queryset'] = self.widget.get_queryset()
-        super(Select2PageSearchField, self).__init__(*args, **kwargs)
+        super().__init__(*args, **kwargs)